not force converting data type of id field to int32 when loading dataset
git-svn-id: https://svn.urbansim.org/src/trunk@19031 1513fa5d-83ce-422e-ae47-6967bb0fb594

+++ opus_core/datasets/dataset.py
@@ -264,7 +264,9 @@ class Dataset(AbstractDataset):
                 else:
                     for id in self._id_names:
                         # Convert all values to int32
-                        self.attribute_boxes[id].set_data(values=self.get_attribute(id).astype(int32))
+                        #self.attribute_boxes[id].set_data(values=self.get_attribute(id).astype(int32))
+                        ## LW: why id field has to be of int32?
+                        self.attribute_boxes[id].set_data(values=self.get_attribute(id).astype('int'))
                         
                     self._create_id_mapping()
